requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
opencv-contrib-python==4.12.0.88
mysql-connector-python==8.2.0
numpy==2.1.0
Pillow==10.2.0
qrcode[pil]==7.4.2
python-dotenv==1.0.0
